Adjust layout of the page, fix various read issues
Many of the options on the page were not explained, and their
purpose wasn't obvious to someone who had never used it. I added
some instructions and tooltips to fix some of these things. For
the code, there were a number of issues with how the dialogue
reader was being accessed. Accessing the dialogue window for the
first time (and subsequent times) should be faster and smoother
now.